Feature #1319
Provides a way to store username and passwords in a clear text file, user.auth, which must be placed in the location of the listenForRemoteConnections.sh script used to run the JavacycServer.  JavacycConnection can be initiallized with a username and password, which will then be authenticated by the server.  Null or empty username and passwords are not allowed.